+++ src/AiAltText.php
@@ -108,10 +108,14 @@ class AiAltText extends Plugin
                 $asset = $event->sender;
 
                 // Only process new assets that are images and if the setting is enabled
+                $isSvg = strtolower($asset->getMimeType()) === 'image/svg+xml';
+                $processSvgs = $this->getSettings()->processSvgs;
+
                 if (
                     $event->isNew
                     && $asset->kind === Asset::KIND_IMAGE
                     && $this->getSettings()->generateForNewAssets
+                    && (!$isSvg || $processSvgs)
                 ) {
                     // Save current site ID
                     $currentSite = Cp::requestedSite();